refactor sbdb funcs from operators subpackage to xlnstorch
