+++ xlnstorch/xlnstorch/operators/__init__.py
@@ -102,13 +102,7 @@ from .internal_lns_ops import (
     lns_avg_pool2d,
     lns_avg_pool3d,
 )
-from .addition_ops import (
-    implement_sbdb,
-    set_default_sbdb_implementation,
-    override_sbdb_implementation,
-    register_xlnsconf_implementation,
-    sbdb,
-)
+from . import addition_ops
 from . import arithmetic_ops
 from . import unary_ops
 from . import comparison_ops
