+++ xlnstorch/xlnstorch/operators/addition_ops.py
@@ -1,156 +1,12 @@
 import torch
-import contextlib
-from typing import Generator, Callable
-from xlnstorch import LNS_ZERO, CSRC_AVAILABLE, lnstensor, format_lnstensor_operands, implements
+from xlnstorch import LNS_ZERO, LNS_ONE, CSRC_AVAILABLE, lnstensor, format_lnstensor_operands, implements, implements_sbdb, sbdb
 from xlnstorch.autograd import LNSFunction
-from xlnstorch.tensor_utils import get_precision_from_base
 from . import (
     lns_add,
     lns_neg,
 )
 
-# SBDB_FUNCS is a dictionary that contains different implementations
-# of the sbdb (Gaussian logarithm) function. Each implementation is
-# registered with a unique key.
-SBDB_FUNCS = {}
-DEFAULT_SBDB_FUNC = ""
-
-def set_default_sbdb_implementation(impl_key: str) -> None:
-    """
-    Set the default implementation for the sbdb function.
-
-    Parameters
-    ----------
-    impl_key : str
-        The key identifying the implementation to be set as default.
-
-    Raises
-    ------
-    ValueError
-        If the specified implementation key is not registered for the sbdb function.
-    """
-    if impl_key not in SBDB_FUNCS:
-        raise ValueError(f"Implementation '{impl_key}' is not registered for the sbdb function.")
-
-    global DEFAULT_SBDB_FUNC
-    DEFAULT_SBDB_FUNC = impl_key
-
-    if CSRC_AVAILABLE:
-        import xlnstorch.csrc
-        xlnstorch.csrc.set_default_sbdb_implementation(impl_key)
-
-@contextlib.contextmanager
-def override_sbdb_implementation(impl_key: str) -> Generator[None, None, None]:
-    """
-    Temporarily override the default sbdb implementation within a context. This
-    allows for testing or using a different implementation without permanently
-    changing the default.
-
-    Parameters
-    ----------
-    impl_key : str
-        The key identifying the new implementation to use as default.
-
-    Yields
-    ------
-    None
-        The function yields control back to the context block.
-    """
-    global DEFAULT_SBDB_FUNC
-    original_default = DEFAULT_SBDB_FUNC
-    set_default_sbdb_implementation(impl_key)
-
-    try:
-        yield
-    finally:
-        DEFAULT_SBDB_FUNC = original_default
-
-def implement_sbdb(key, default=False):
-    """
-    A decorator to register a custom sbdb implementation. This will
-    be used to compute/approximate the Gaussian logarithms for the
-    addition and subtraction operations in the logarithmic domain. See
-
-    https://en.wikipedia.org/wiki/Logarithmic_number_system
-    https://en.wikipedia.org/wiki/Gaussian_logarithm
-
-    Parameters
-    ----------
-    key : str
-        The key to register the sbdb function under. This should be
-        unique across all sbdb implementations.
-    default : bool, optional
-        If True, this sbdb function will be set as the default sbdb
-        implementation. If multiple sbdb functions are registered
-        with `default=True`, the last one registered will be used as
-        the default. Defaults to False.
-
-    Raises
-    ------
-    ValueError
-        If an sbdb function with the given key is already registered.
-    """
-    def decorator(func):
-        function_key = key or func.__name__
-
-        if function_key in SBDB_FUNCS:
-            raise ValueError(f"sbdb function with key '{function_key}' is already implemented.")
-        SBDB_FUNCS[function_key] = func
-
-        if default:
-            global DEFAULT_SBDB_FUNC
-            DEFAULT_SBDB_FUNC = function_key
-
-        return func
-    return decorator
-
-def register_xlnsconf_implementation(xlns_function: Callable, impl_key: str) -> None:
-    """
-    """
-    if impl_key in SBDB_FUNCS:
-        raise ValueError(f"Implementation '{impl_key}' is already registered for the sbdb function.")
-
-    def wrapper_sbdb(z, s, base):
-        precision = get_precision_from_base(base)
-        z_np = z.numpy()
-        s_np = s.numpy()
-
-        xlns_result = xlns_function(z_np, s_np, B=base.item(), F=precision)
-        return torch.tensor(xlns_result, dtype=torch.int64)
-
-    SBDB_FUNCS[impl_key] = wrapper_sbdb
-
-def sbdb(z, s, base):
-    """
-    Computes the Gaussian logarithm for the given inputs z and s.
-
-    Parameters
-    ----------
-    z : torch.Tensor
-        The negation of the absolute difference between the two operands
-        in the logarithmic domain.
-    s : torch.Tensor
-        The sign difference between the two operands in the logarithmic
-        domain.
-    base : torch.Tensor
-        The base of the operands. Required for certain sbdb implementations.
-
-    Returns
-    -------
-    torch.Tensor
-        The result of the Gaussian logarithm computation.
-
-    Raises
-    ------
-    ValueError
-        If no default sbdb function is implemented.
-    """
-    if DEFAULT_SBDB_FUNC not in SBDB_FUNCS:
-        raise ValueError(f"No default sbdb function implemented.")
-
-    return SBDB_FUNCS[DEFAULT_SBDB_FUNC](z, s, base)
-
-@implement_sbdb('ideal', default=True)
+@implements_sbdb('ideal', default=True)
 def sbdb_ideal(z, s, base):
     """
     Ideal implementation of the sbdb function that directly computes:
@@ -252,6 +108,64 @@ def sub(x, y, *, alpha=1, out=None):
         y = torch.mul(y, alpha)
 
     result = LNSSubFunction.apply(x, y, x.base)
+
+    if out is not None:
+        return out._inplace_copy(result)
+
+    return lnstensor(result, from_lns=True, b=x.base)
+
+class LNSSumFunction(LNSFunction):
+    """
+    We use the addition operation to compute the sum.
+
+    Gradients are computed as follows:
+    d/dx(sum(x)) = 1
+    """
+
+    @staticmethod
+    def forward(x, base, dim=None, keepdim=False):
+        if dim is None:
+            flat = x.reshape(-1)
+            out = flat[0]
+            for i in range(1, flat.numel()):
+                out = lns_add(out, flat[i], base)
+            if keepdim:
+                out = out.reshape([1] * x.dim())
+            return out
+
+        red_dims = (dim,) if isinstance(dim, int) else tuple(dim)
+        red_dims = tuple(sorted(d % x.dim() for d in red_dims))
+
+        permute_order = [d for d in range(x.dim()) if d not in red_dims] + list(red_dims)
+        transposed = x.permute(*permute_order)
+
+        outer_shape = transposed.shape[:-len(red_dims)]
+        transposed = transposed.reshape(*outer_shape, -1)
+
+        out = transposed[..., 0]
+        for i in range(1, transposed.shape[-1]):
+            out = lns_add(out, transposed[..., i], base)
+
+        if keepdim:
+            for d in red_dims:
+                out = out.unsqueeze(d)
+
+        return out
+
+    @staticmethod
+    def setup_context(ctx, inputs, output):
+        x, _, _, _ = inputs
+        ctx.save_for_backward(x)
+
+    @staticmethod
+    def backward(ctx, grad_output):
+        x, = ctx.saved_tensors
+        return torch.full_like(x, LNS_ONE.item()), None, None, None
+
+@implements(torch.sum, LNSSumFunction.forward, "default", default=not CSRC_AVAILABLE)
+def sum(x, dim=None, keepdim=False, *, out=None):
+
+    result = LNSSumFunction.apply(x, x.base, dim, keepdim)
 
     if out is not None:
         return out._inplace_copy(result)
